refactor and move validation in utils
