+++ packages/client/src/modules/play/Personalization/utils/formValidation.ts
@@ -1,0 +1,31 @@
+const compare = (watch: any, condition: any) => {
+  if (condition.operator === ">") {
+    return watch(condition.question) > condition.value;
+  }
+  if (condition.operator === "<") {
+    return watch(condition.question) < condition.value;
+  }
+  if (condition.operator === "=") {
+    return watch(condition.question) === condition.value;
+  }
+  return false;
+};
+
+export const fulfillsConditions = (watch: any, question: any) => {
+  if (!question.conditions) {
+    return true;
+  }
+  return question.conditions.every((condition: any) =>
+    compare(watch, condition)
+  );
+};
+
+export const isSectionValid = (
+  formValues: any,
+  watch: any,
+  sectionName: string
+) => {
+  return formValues
+    .filter((question: any) => question.type === sectionName)
+    .every((question: any) => watch(question.name) !== undefined);
+};
